Fix 11 clippy warnings for CI (bevy-server)
- contains_key instead of get().is_none()
- clamp instead of min/max chain
- div_ceil instead of manual (n+7)/8
- struct init instead of field reassign
- allow(too_many_arguments) for complex functions
- inner doc comments (//!) in ffi.rs
- remove unnecessary u32 cast

+++ bevy-server/src/loot.rs
@@ -561,7 +561,7 @@ pub fn generate_loot(seed: u64, config: &LootConfig) -> Vec<LootDrop> {
                     .iter()
                     .map(|(k, v)| (k.clone(), *v))
                     .collect(),
-                gold_value: (50 * config.loot_tier) as u32,
+                gold_value: 50 * config.loot_tier,
             });
         }
     }
